fix : stress correctly updated + better formulas + issue where score 1.0 but no possibilities given on decision feat : damage system --> to upgrade idk how to calculate damage when fire from on point

+++ pgl/src/main/java/fr/cy/model/agent/behaviour/properties/AgentPhysicalProperties.java
@@ -7,14 +7,14 @@ public class AgentPhysicalProperties implements Serializable {
     /** Maximum speed of the agent, in units per time step */
     private double maxSpeed;
 
-    private int maxHealth;
+    private double maxHealth;
     /** Current health level of the agent, between 0 and 100 */
-    private int health;
+    private double health;
 
     /** Surface area taken by the agent, used to calculate crowding effects */
     private double surfaceAreaTakenByAgent = 0.5;
 
-    public AgentPhysicalProperties(double maxSpeed, int maxHealth, int health, double surfaceAreaTakenByAgent) {
+    public AgentPhysicalProperties(double maxSpeed, double maxHealth, double health, double surfaceAreaTakenByAgent) {
         this.maxSpeed = maxSpeed;
         this.maxHealth = maxHealth;
         if (health < 0 || health > maxHealth) {
@@ -32,26 +32,26 @@ public class AgentPhysicalProperties implements Serializable {
         return surfaceAreaTakenByAgent;
     }
 
-    public void setHealth(int health) {
+    public void setHealth(double health) {
         if (health < 0 || health > maxHealth) {
             throw new IllegalArgumentException("Health must be between 0 and maxHealth");
         }
         this.health = health;
     }
 
-    public int getHealth() {
+    public double getHealth() {
         return health;
     }
 
-    public int getMaxHealth() {
+    public double getMaxHealth() {
         return maxHealth;
     }
 
-    public void decreaseHealth(int amount) {
+    public void decreaseHealth(double amount) {
         setHealth(Math.max(0, getHealth() - amount));
     }
 
-    public void restoreHealth(int amount) {
+    public void restoreHealth(double amount) {
         setHealth(Math.min(maxHealth, getHealth() + amount));
     }
